fix case where bondy.conf.template does not exist

+++ start.sh
@@ -2,10 +2,13 @@
 
 set -e
 
-echo "Replacing os environment variables in bondy.conf"
-cp /bondy/etc/bondy.conf.template /tmp/bondy.conf.template
-/bondy/etc/replace-env-vars -i /tmp/bondy.conf.template -o /tmp/bondy.conf
-cp /tmp/bondy.conf /bondy/etc/bondy.conf
+USER_CONF=/bondy/etc/bondy.conf.template
+if test -f "$USER_CONF"; then
+    echo "Replacing os environment variables in $USER_CONF"
+    cp /bondy/etc/bondy.conf.template /tmp/bondy.conf.template
+    /bondy/etc/replace-env-vars -i /tmp/bondy.conf.template -o /tmp/bondy.conf
+    cp /tmp/bondy.conf /bondy/etc/bondy.conf
+fi
 export RUNNER_ETC_DIR=/bondy/etc
 export RUNNER_GEN_DIR=$RUNNER_ETC_DIR
 export RUNNER_LOG_DIR=/bondy/log
